unpin eslint-config-next for Next.js framework tests (#45)
TLDR; This change changes our dev dependencies to use the latest version of `eslint-config-next`.

We had a floating dependency on Next.js, which would grab the latest version to test against. However, we had pinned our `eslint-config-next` to a particular version.

It turns out that particular version pulled in a Next.js dependency that contained a security vulnerability.